tried fixing saving/loading again

+++ src/plugin/Main.java
@@ -103,7 +103,8 @@ public class Main extends JavaPlugin implements Listener {
             // actual saving takes place here
             for (ClaimedChunk chunk : claimedChunks) {
                 double[] coords = chunk.getCoordinates();
-                saveWriter.write(coords[0] + "" + coords[1] + ".txt" + "\n");
+
+                saveWriter.write((int)coords[0] + "" + (int)coords[1] + ".txt" + "\n");
             }
 
             saveWriter.close();
